requirements.txt
pandas
google-generativeai
python-dotenv
PyPDF2
rich